Strip out the V1 UI detritus
